重整项目结构
