+++ src/main/java/cn/gdufe/service/StudentService.java
@@ -1,44 +1,32 @@
 package cn.gdufe.service;
 
-import cn.gdufe.domain.PageBean;
 import cn.gdufe.domain.Student;
-
-import java.util.Map;
 
 public interface StudentService {
 
     /**
      * 添加学生
-     * @param student
+     * @param student 学生对象
      */
     void addStudent(Student student);
 
     /**
      * 通过id查询学生
-     * @param sid
-     * @return
+     * @param sid 学生id
+     * @return 学生对象
      */
     Student findStudent(String sid);
 
 
     /**
      * 删除学生
-     * @param sid
+     * @param sid 学生id
      */
     void deleteStudent(String sid);
 
     /**
-     * 分页查询学生
-     * @param currentPage 当前页码
-     * @param rows
-     * @param condition
-     * @return
-     */
-    PageBean<Student> findStudentByPage(String currentPage, String rows, Map<String, String[]> condition);
-
-    /**
      * 更新学生信息
-     * @param student
+     * @param student 学生对象
      */
     void updateStudent(Student student);
 }
