+++ src/main/java/cn/gdufe/service/UserService.java
@@ -10,27 +10,27 @@ public interface UserService {
 
     /**
      * 登录
-     * @param loginUser
-     * @return
+     * @param loginUser 请求登录的用户对象
+     * @return 查询的用户对象
      */
     User login(User loginUser);
 
     /**
      * 通过用户名查询用户
-     * @param username
-     * @return
+     * @param username 用户名
+     * @return 用户对象
      */
     User findUser(String username);
 
     /**
      * 添加用户
-     * @param user
+     * @param user 用户对象
      */
     void addUser(User user);
 
     /**
      * 更新用户信息
-     * @param user
+     * @param user 用户对象
      */
     void updateUser(User user);
 }
