+++ src/main/java/cn/gdufe/service/impl/TeacherServiceImpl.java
@@ -2,12 +2,8 @@ package cn.gdufe.service.impl;
 
 import cn.gdufe.dao.TeacherDao;
 import cn.gdufe.dao.impl.TeacherDaoImpl;
-import cn.gdufe.domain.PageBean;
 import cn.gdufe.domain.Teacher;
 import cn.gdufe.service.TeacherService;
-
-import java.util.List;
-import java.util.Map;
 
 public class TeacherServiceImpl implements TeacherService {
     private final TeacherDao dao = new TeacherDaoImpl();
@@ -25,32 +21,6 @@ public class TeacherServiceImpl implements TeacherService {
     @Override
     public void deleteTeacher(String tid) {
         dao.deleteTeacher(tid);
-    }
-
-    @Override
-    public PageBean<Teacher> findTeacherByPage(String _currentPage, String _rows, Map<String, String[]> condition) {
-        int currentPage=Integer.parseInt(_currentPage);
-        int rows=Integer.parseInt(_rows);
-        if (currentPage<=0){
-            currentPage=1;
-        }
-        //1.创建空PageBean对象
-        PageBean<Teacher> teacherPageBean = new PageBean<>();
-        //2.设置参数
-        teacherPageBean.setCurrentPage(currentPage);
-        teacherPageBean.setRows(rows);
-        //3.调用dao查询总记录数
-        int totalCount=dao.findTotalCount(condition);
-        teacherPageBean.setTotalPage(totalCount);
-        //4.调用dao查询List集合
-        //计算开始的记录索引
-        int start=(currentPage-1)*rows;
-        List<Teacher> list=dao.findByPage(start,rows,condition);
-        teacherPageBean.setList(list);
-        //5.计算总页码
-        int totalPage=(totalCount%rows)==0?(totalCount/rows):(totalCount/rows)+1;
-        teacherPageBean.setTotalPage(totalPage);
-        return teacherPageBean;
     }
 
     @Override
