+++ src/main/java/cn/gdufe/web/servlet/course/UpdateClassServlet.java
@@ -1,8 +1,10 @@
-package cn.gdufe.web.servlet;
+package cn.gdufe.web.servlet.course;
 
-import cn.gdufe.domain.User;
-import cn.gdufe.service.impl.UserServiceImpl;
+import cn.gdufe.domain.Class;
+import cn.gdufe.service.ClassService;
+import cn.gdufe.service.impl.ClassServiceImpl;
 import org.apache.commons.beanutils.BeanUtils;
+
 import javax.servlet.ServletException;
 import javax.servlet.annotation.WebServlet;
 import javax.servlet.http.HttpServlet;
@@ -12,29 +14,28 @@ import java.io.IOException;
 import java.lang.reflect.InvocationTargetException;
 import java.util.Map;
 
-@WebServlet("/updateUserServlet")
-public class UpdateUserServlet extends HttpServlet {
+@WebServlet("/updateClassServlet")
+public class UpdateClassServlet extends HttpServlet {
     @Override
     protected void doPost(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
         //1.设置编码
         request.setCharacterEncoding("utf-8");
-        //2.获取map
+
+        //2.获取参数
         Map<String, String[]> map = request.getParameterMap();
         //3.封装对象
-        User user = new User();
+        Class c = new Class();
         try {
-            BeanUtils.populate(user,map);
+            BeanUtils.populate(c,map);
         } catch (IllegalAccessException | InvocationTargetException e) {
             e.printStackTrace();
         }
+        //4.调用service
+        ClassService service = new ClassServiceImpl();
+        service.updateClass(c);
 
-        //4.调用Service修改
-        UserServiceImpl service = new UserServiceImpl();
-        service.updateUser(user);
-
-        //5.跳转到查询所有Servlet
-        response.sendRedirect(request.getContextPath()+"/menu.jsp");
-
+        //5.跳转查询页面
+        response.sendRedirect(request.getContextPath()+"/findClassByPageServlet");
     }
 
     @Override
