+++ src/main/java/cn/gdufe/web/servlet/student/AddStudentServlet.java
@@ -1,4 +1,4 @@
-package cn.gdufe.web.servlet;
+package cn.gdufe.web.servlet.student;
 
 
 import cn.gdufe.domain.Student;
