+++ src/main/java/cn/gdufe/web/servlet/student/DelStudentServlet.java
@@ -1,8 +1,7 @@
-package cn.gdufe.web.servlet;
+package cn.gdufe.web.servlet.student;
 
-import cn.gdufe.service.ClassService;
-import cn.gdufe.service.impl.ClassServiceImpl;
-
+import cn.gdufe.service.StudentService;
+import cn.gdufe.service.impl.StudentServiceImpl;
 import javax.servlet.ServletException;
 import javax.servlet.annotation.WebServlet;
 import javax.servlet.http.HttpServlet;
@@ -10,15 +9,15 @@ import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import java.io.IOException;
 
-@WebServlet("/delClassServlet")
-public class DelClassServlet extends HttpServlet {
+@WebServlet("/delStudentServlet")
+public class DelStudentServlet extends HttpServlet {
     @Override
     protected void doPost(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
         //1.获取id
-        String classId = request.getParameter("classId");
+        String sid = request.getParameter("sid");
         //2.调用service删除
-        ClassService service = new ClassServiceImpl();
-        service.deleteClass(classId);
+        StudentService service = new StudentServiceImpl();
+        service.deleteStudent(sid);
 
         //3.跳转到查询所有Servlet
         response.sendRedirect(request.getContextPath()+"/menu.jsp");
