+++ src/main/java/cn/gdufe/web/servlet/student/FindStudentServlet.java
@@ -1,7 +1,9 @@
-package cn.gdufe.web.servlet;
+package cn.gdufe.web.servlet.student;
 
+import cn.gdufe.domain.Student;
 import cn.gdufe.service.StudentService;
 import cn.gdufe.service.impl.StudentServiceImpl;
+
 import javax.servlet.ServletException;
 import javax.servlet.annotation.WebServlet;
 import javax.servlet.http.HttpServlet;
@@ -9,18 +11,23 @@ import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import java.io.IOException;
 
-@WebServlet("/delStudentServlet")
-public class DelStudentServlet extends HttpServlet {
+@WebServlet("/findStudentServlet")
+public class FindStudentServlet extends HttpServlet {
     @Override
     protected void doPost(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
-        //1.获取id
-        String sid = request.getParameter("sid");
-        //2.调用service删除
-        StudentService service = new StudentServiceImpl();
-        service.deleteStudent(sid);
+        //1.设置编码
+        request.setCharacterEncoding("utf-8");
 
-        //3.跳转到查询所有Servlet
-        response.sendRedirect(request.getContextPath()+"/menu.jsp");
+        //2.获取参数
+        String sid = request.getParameter("sid");
+        StudentService service = new StudentServiceImpl();
+        Student student = service.findStudent(sid);
+
+        //3.将user存入request
+        request.setAttribute("student",student);
+
+        //4.转发到update.jsp
+        request.getRequestDispatcher("/edit_Student.jsp").forward(request,response);
     }
 
     @Override
