+++ src/main/webapp/index.jsp
@@ -1,16 +1,16 @@
 <%--
   Created by IntelliJ IDEA.
   User: A
-  Date: 2022/5/4
-  Time: 18:43
+  Date: 2022/5/7
+  Time: 16:31
   To change this template use File | Settings | File Templates.
 --%>
-<%@ page contentType="text/html;charset=UTF-8" language="java" %>
+<%@ page contentType="text/html;charset=UTF-8"%>
 <html>
 <head>
     <title>Title</title>
 </head>
 <body>
-<h1>hello!</h1>
+<h1>hello</h1>
 </body>
 </html>
